Compare to Vim’s built-in Python support
Made it clear that Snake doesn’t add the Python integration, but rather improves the integration so it is actually easy enough to write plugins with.

Mentioned the requirement for `+python` in “how do I get it” so people know the prerequisites for the software. This should help, for example, people running gvim on Windows, which does not include `+python`. I wrote the command `:version` instead of `:echo has('python')` to make the output easier to understand for those who don’t know that 0 and 1 are booleans in Vim.

Fixes https://github.com/amoffat/snake/issues/5