Fix problem with insert_mode key_maps
This fixes the problem I see when I attempt to write an insert mode key
map in my .vimrc.py.

That problem is that the text of the map (
:python3 snake.dispatch_mapped_function(140292554737728) )

is inserted into the buffer instead of leaving insert mode, running the
text of the map, and then returning to insert mode.